// types/Empleado/IVerEmpleado.ts
export interface IVerEmpleado {
  id: string;
  nombreCompleto: string;
  nombre: string;
  apellidos: string;
  genero: string;
  estadoCivil: string;
  tipoDocumento: string;
  numeroDocumento: string;
  fechaNacimiento: string;
  edad: number;
  correo: string;
  telefonoMovil: string;
  fotoUrl: string | null;
  nacionalidad: string;
  direccion: string | null;
  departamento: string | null;
  provincia: string | null;
  distrito: string | null;
  contactoEmergenciaNombre: string | null;
  contactoEmergenciaTelefono: string | null;
  contactoEmergenciaParentesco: string | null;
  bancoNombre: string | null;
  numeroCuentaBancaria: string | null;
  tipoCuenta: string | null;
  cci: string | null;
  numeroEssalud: string | null;
  sistemaPensiones: string | null;
  cuspp: string | null;
  nivelEducativo: string | null;
  profesionOficio: string | null;
  codigoEmpleado: string;
  cargoActual: string;
  salarioActual: number;
  fechaIngresoActual: string;
  tipoContrato: string | null;
  tipoJornada: string | null;
  isActive: boolean;
}